update direction of images
I noticed the parallax images would move ahead of the page scroll speed. This seemed off to me because parallax implies the image is recessed and moving at a slower speed in relation to our perspective.

in the onScroll() function I modified the direction the images would move (line 67), as well as bit of extra math on the transformY variable (line 65) to account for the new offset of the images.

the result is that the images now 'lag' slightly behind the scroll speed of the page, creating the illusion that they are further away.